pyproject.toml
[project]
name = "ais"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-snowflake>=1.11.3",
    "fastapi>=0.135.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "snowflake-connector-python>=4.3.0",
    "uvicorn[standard]>=0.41.0",
]
